Candidate resources moved into separate file
