+++ webservices/resources.py
@@ -4,6 +4,7 @@ import logging
 import time
 
 from flask.ext import restful
+from flask.ext.restful import reqparse
 import htsql
 from psycopg2._range import DateTimeRange
 import sqlalchemy as sa
@@ -44,6 +45,14 @@ def default_year():
     year = datetime.now().year
     years = [str(y) for y in range(year, year-4, -1)]
     return ','.join(years)
+
+
+def natural_number(n):
+    result = int(n)
+    if result < 1:
+        raise reqparse.ArgumentTypeError(
+            'Must be a number greater than or equal to 1')
+    return result
 
 
 class SingleResource(restful.Resource, FindFieldsMixin):
